// src/types/enums/reportReasonEnum.ts
export enum ReasonEnum {
  VIOLENCE = "violence",
  REGULATED_GOODS_AND_ACTIVITIES = "regulated-goods",
  HATE_OR_HARASSMENT = "hate-harassment",
  FRAUD = "fraud",
  PRETENDING_TO_BE_SOMEONE_ELSE = "pretending",
  INTELLECTUAL_PROPERTY_VIOLATION = "ip-violation",
  SPAM = "spam",
  OTHER = "other",
}